deduplicate `RangeSet`
